remove circular dependencies in core
