+++ packages/core/rollup.config.mjs
@@ -22,6 +22,7 @@ const globals = {
   'flat': 'flat',
   'papaparse': 'papaparse',
   'redux-persist': 'reduxPersist',
+  'cookies-next': 'cookies-next'
 };
 
 const config = [
