+++ packages/frontend/package.json
@@ -61,7 +61,7 @@
     "@mantine/notifications": "^6.0.21",
     "@mdx-js/loader": "^2.1.5",
     "@mdx-js/mdx": "^2.1.5",
-    "@next/mdx": "^14.1.0",
+    "@next/mdx": "^14.1.1",
     "@reduxjs/toolkit": "^1.9.5",
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tailwindcss/forms": "^0.5.3",
@@ -85,7 +85,7 @@
     "jsonpath-plus": "^8.0.0",
     "mantine-react-table": "^1.3.3",
     "minisearch": "^6.3.0",
-    "next": "^14.1.0",
+    "next": "^14.1.1",
     "next-compose-plugins": "^2.2.1",
     "next-images": "^1.8.5",
     "pluralize": "^8.0.0",
