+++ packages/sampleCommons/package.json
@@ -20,7 +20,7 @@
     "@gen3/frontend": "file:../frontend",
     "@mdx-js/loader": "^3.0.0",
     "@mdx-js/react": "^3.0.0",
-    "@next/mdx": "^14.1.0",
+    "@next/mdx": "^14.1.1",
     "@types/mdx": "^2.0.11",
     "cookies-next": "^4.0.0",
     "jose": "^4.15.4"
@@ -29,7 +29,7 @@
     "@gen3/toolsff": "file:../tools",
     "@types/react": "^18.2.21",
     "@types/react-dom": "^18.2.7",
-    "next": "^14.1.0",
+    "next": "^14.1.1",
     "react": "^18.2.0",
     "react-dom": "18.2.0",
     "tailwindcss": "^3.4.1",
